Improved reporting of line numbers for syntax errors when using YAP or SWI-Prolog as the back-end compilers.
Added entity property lines(Start, End) for entities compiled and loaded from source files. This property is only available when using SWI-Prolog or YAP as the back-end compilers.

Added a customized version of the read_term/2 predicate to all config files that returns the line where a term starts for back-end Prolog compilers whose implementation of the read_term/2 predicate supports this information.

git-svn-id: http://svn.logtalk.org/logtalk/trunk@4517 ac9c905b-388a-4ddf-9b6a-4063a022bc57