Added an optimization and security compiler option, "complements", that allows objects to be compiled with support for complementing categories either enabled or disabled. The default value (set in the config files) is "off" (disabled).
Added an optimization and security compiler option, "dynamic_declarations", that allows objects to be compiled with support for dynamic declaration of predicates (using the built-in database methods asserta/1 and assertz/1) either enabled or disabled. The default value (set in the config files) is "off" (disabled).

Updated the "complements" example to use the new "complements" compiler option. Update the "dynpred" example to use the new "dynamic_declarations" compiler option.

git-svn-id: http://svn.logtalk.org/logtalk/trunk@4524 ac9c905b-388a-4ddf-9b6a-4063a022bc57