Updated version number to 2.39.2 in preparation for the next stable release.
Updated the Logtalk compiler in order to reject use_module/2 directives where the first argument is not an atom (assumed to be the module name). This change results in using the config files to find the module name from the module file name when the first argument of the use_module/2 directive is a compound term (as usual when using library notation).

Updated the ECLiPSe, SICStus Prolog, SWI-Prolog, and YAP config files in order to preprocess the use_module/2 directive, converting the module file name into the module name.

git-svn-id: http://svn.logtalk.org/logtalk/trunk@5405 ac9c905b-388a-4ddf-9b6a-4063a022bc57